+++ lib/src/dart_formatter.dart
@@ -35,7 +35,7 @@ final RegExp _widthCommentPattern = RegExp(r'^// dart format width=(\d+)$');
 final class DartFormatter {
   /// The latest Dart language version that can be parsed and formatted by this
   /// version of the formatter.
-  static final latestLanguageVersion = Version(3, 11, 0);
+  static final latestLanguageVersion = Version(3, 12, 0);
 
   /// The latest Dart language version that will be formatted using the older
   /// "short" style.