+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 ## 3.1.6-wip
 
+* Format extension type representation clauses the same way primary constructor
+  formal parameter lists are formatted. This rarely makes a difference but
+  produces better formatting when the representation type is long and there are
+  other clauses on the extension type, as in:
+
+  ```dart
+  // Before:
+  extension type JSExportedDartFunction._(
+    JSExportedDartFunctionRepType _jsExportedDartFunction
+  )
+      implements JSFunction {}
+
+  // After:
+  extension type JSExportedDartFunction._(
+    JSExportedDartFunctionRepType _jsExportedDartFunction
+  ) implements JSFunction {}
+  ```
+
+  This change is *not* language versioned. (The old style is always worse, and
+  continuing to support it would add complexity to the formatter.)
+
 * Require `sdk: ^3.10.0`.
 * Stop using experiment flags for features released in 3.10.